tests: extract Source boundary into test_source.py
Source factories (from_file, from_arxiv, from_json, from_uri) are a
distinct "user input enters" surface with their own validation and error
type (SourceError).  They were living in test_pipeline.py as a catch-all.

- Move 4 Source construction/validation tests to new test_source.py
- Keep test_source_from_json_is_sent_as_inline_content in pipeline
  (exercises run() end-to-end, not just Source construction)
- Update TESTING.md boundary table and test placement guide
- Update AGENTS.md test files table
